Commit plug-in work
