--- plugins/my-own-plugin-test/Render.svelte
@@ -1,3 +1,0 @@
-
-
-Hello world!
