+++ plugins/vimeo-player/InterruptionForm.svelte
@@ -1,7 +1,11 @@
 <script>
   import FormField from "$lib/Internals/FormFields/FormField.svelte";
   import NumberInput from "$lib/Internals/FormFields/NumberInput.svelte";
-  export let data = {};
+  export let data;
+
+  $: if (!data) {
+    data = {};
+  }
 </script>
 
 <FormField title="Tijd in video" forId="video_time" labelPosition="top">
