+++ postcss.config.cjs
@@ -1,4 +1,3 @@
-
 const tailwindcss = require("tailwindcss");
 const autoprefixer = require("autoprefixer");
 const cssnano = require("cssnano");
@@ -8,18 +7,17 @@ const dev = mode === "development";
 
 module.exports = {
   plugins: [
+    //Some plugins, like tailwindcss/nesting, need to run before Tailwind,
+    tailwindcss(),
+    //But others, like autoprefixer, need to run after,
     // Some plugins, like postcss-nested, need to run before Tailwind
-
-    tailwindcss,
 
     // But others, like autoprefixer, need to run after
 
     autoprefixer,
-
-    !dev && cssnano({
-      preset: "default",
-    }),
-
-
+    !dev &&
+      cssnano({
+        preset: "default",
+      }),
   ],
 };
