+++ src/app.css
@@ -1,5 +1,7 @@
 /* Write your global styles here, in PostCSS syntax */
 @tailwind base;
+@tailwind base;
 @tailwind components;
 @tailwind utilities;
-
+@tailwind components;
+@tailwind utilities;
